feat: SSGではなくSSRでの実装に変更
