--- src/app/users/show/page.tsx
@@ -1,5 +1,0 @@
-import { UserShow } from './UserShow'
-
-export default function Page() {
-  return <UserShow />
-}
